chore: remove old btc witnessing
