+++ engine/src/btc/mod.rs
@@ -2,8 +2,6 @@ use self::rpc::BtcRpcApi;
 
 pub mod retry_rpc;
 pub mod rpc;
-pub mod witnesser;
-pub mod witnessing;
 
 use anyhow::Result;
 use bitcoin::Txid;
